ProcessorState.uptime is UInt64 now.

+++ TurtleTTL/Simulator/ComputerPeripherals.swift
@@ -13,7 +13,7 @@ public class ComputerPeripherals: NSObject {
     public var bus = Register()
     public var registerX = Register()
     public var registerY = Register()
-    public private(set) var uptime: Int = 0
+    public private(set) var uptime: UInt64 = 0
     public var peripherals: [ComputerPeripheral] = [ComputerPeripheral(),
                                                     ComputerPeripheral(),
                                                     ComputerPeripheral(),
@@ -105,7 +105,7 @@ public class ComputerPeripherals: NSObject {
         return peripherals[6] as! SerialInterfacePeripheral
     }
     
-    public func catchUp(uptime desiredUptime: Int) {
+    public func catchUp(uptime desiredUptime: UInt64) {
         logger?.append("Peripherals are catching up to CPU now. -- " +
                        "peripheral uptime: \(uptime) ; " +
                        "CPU uptime: \(desiredUptime)")
